+++ src/osgPlugins/flt/flt.h
@@ -65,8 +65,8 @@ typedef signed char     int8;
 typedef unsigned char   uint8;
 typedef signed short    int16;
 typedef unsigned short  uint16;
-typedef signed long     int32;
-typedef unsigned long   uint32;
+typedef signed int     	int32;	
+typedef unsigned int   	uint32;	
 typedef float           float32;
 typedef double          float64;
